Add Search Bar for Stocks
* Create Stock Details page
* Move Edit Stock dialog to separate component and reuse in Stock Details page
* Various smaller improvements
