+++ packages/rating-tracker-frontend/src/layouts/SidebarLayout/Header/Buttons/Search/index.tsx
@@ -1,33 +1,43 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { forwardRef, Ref, useState, ReactElement, ChangeEvent } from "react";
+import React, {
+  forwardRef,
+  Ref,
+  useState,
+  ReactElement,
+  ChangeEvent,
+} from "react";
 import {
   Avatar,
-  Link,
   Box,
-  Button,
   Divider,
   IconButton,
   InputAdornment,
-  lighten,
   List,
   ListItem,
   ListItemAvatar,
   TextField,
-  Theme,
   Tooltip,
   Typography,
   Dialog,
   DialogContent,
   DialogTitle,
   Slide,
-  Hidden,
+  Skeleton,
 } from "@mui/material";
-import { styled } from "@mui/material/styles";
+import { styled, useTheme } from "@mui/material/styles";
 import { TransitionProps } from "@mui/material/transitions";
 import SearchIcon from "@mui/icons-material/Search";
-import FindInPageIcon from "@mui/icons-material/FindInPage";
 
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
+import axios from "axios";
+import {
+  baseUrl,
+  logoEndpoint,
+  stockAPI,
+  stockListEndpoint,
+} from "../../../../../endpoints";
+import { emojiFlag, Stock } from "rating-tracker-commons";
+import useNotification from "../../../../../helpers/useNotification";
+import SectorIcon from "../../../../../components/SectorIcon/";
 
 const Transition = forwardRef(function Transition(
   props: TransitionProps & { children: ReactElement<any, any> },
@@ -48,16 +58,6 @@ const DialogWrapper = styled(Dialog)(
 `
 );
 
-const SearchInputWrapper = styled(TextField)(
-  ({ theme }) => `
-    background: ${theme.colors.alpha.white[100]};
-
-    .MuiInputBase-input {
-        font-size: ${theme.typography.pxToRem(17)};
-    }
-`
-);
-
 const DialogTitleWrapper = styled(DialogTitle)(
   ({ theme }) => `
     background: ${theme.colors.alpha.black[5]};
@@ -65,19 +65,36 @@ const DialogTitleWrapper = styled(DialogTitle)(
 `
 );
 
+const HoverListItem = styled(ListItem)(
+  ({ theme }) => `
+  &.MuiListItem-root {
+    padding-top: 13.5px;
+    padding-bottom: 13.5px;
+  }
+
+  &.MuiListItem-root:hover {
+    background-color: ${theme.colors.alpha.black[5]};
+  }
+    `
+);
+
 function HeaderSearch() {
-  const [openSearchResults, setOpenSearchResults] = useState(false);
   const [searchValue, setSearchValue] = useState("");
+  const [stocks, setStocks] = useState<Stock[]>([]);
+  const [count, setCount] = useState<number>(0);
+  const [stocksFinal, setStocksFinal] = useState<boolean>(false);
+
+  const { setNotification } = useNotification();
+  const theme = useTheme();
 
   const handleSearchChange = (event: ChangeEvent<HTMLInputElement>): void => {
     setSearchValue(event.target.value);
 
     if (event.target.value) {
-      if (!openSearchResults) {
-        setOpenSearchResults(true);
-      }
+      getStocks(event.target.value);
     } else {
-      setOpenSearchResults(false);
+      setStocks([]);
+      setCount(0);
     }
   };
 
@@ -88,12 +105,43 @@ function HeaderSearch() {
   };
 
   const handleClose = () => {
+    setSearchValue("");
+    setStocks([]);
+    setCount(0);
     setOpen(false);
+  };
+
+  const getStocks = (currentSearchValue: string) => {
+    setStocksFinal(false);
+    axios
+      .get(baseUrl + stockAPI + stockListEndpoint, {
+        params: {
+          sortBy: "name",
+          name: currentSearchValue,
+        },
+      })
+      .then((res) => {
+        setStocks(res.data.stocks.map((stock: any) => new Stock(stock)));
+        setCount(res.data.count);
+      })
+      .catch((e) => {
+        setNotification({
+          severity: "error",
+          title: "Error while fetching stock information",
+          message:
+            e.response?.status && e.response?.data?.message
+              ? `${e.response.status}: ${e.response.data.message}`
+              : e.message ?? "No additional information available.",
+        });
+        setStocks([]);
+        setCount(0);
+      })
+      .finally(() => setStocksFinal(true));
   };
 
   return (
     <>
-      {/* <Tooltip arrow title="Search">
+      <Tooltip arrow title="Search">
         <IconButton color="primary" onClick={handleClickOpen}>
           <SearchIcon />
         </IconButton>
@@ -103,176 +151,155 @@ function HeaderSearch() {
         open={open}
         TransitionComponent={Transition}
         keepMounted
-        maxWidth="md"
+        maxWidth="xs"
         fullWidth
         scroll="paper"
         onClose={handleClose}
       >
         <DialogTitleWrapper>
-          <SearchInputWrapper
+          <TextField
             value={searchValue}
-            autoFocus={true}
+            inputRef={(input) => {
+              if (input !== null && open) {
+                input.focus();
+              }
+            }}
+            disabled={!open}
             onChange={handleSearchChange}
             InputProps={{
               startAdornment: (
                 <InputAdornment position="start">
                   <SearchIcon />
                 </InputAdornment>
-              )
+              ),
             }}
-            placeholder="Search terms here..."
+            placeholder="Enter a stock name or symbol…"
             fullWidth
             label="Search"
           />
         </DialogTitleWrapper>
         <Divider />
-
-        {openSearchResults && (
-          <DialogContent>
-            <Box
-              sx={{ pt: 0, pb: 1 }}
-              display="flex"
-              justifyContent="space-between"
-            >
-              <Typography variant="body2" component="span">
-                Search results for{' '}
-                <Typography
-                  sx={{ fontWeight: 'bold' }}
-                  variant="body1"
-                  component="span"
+        {(searchValue || count > 0) && (
+          <DialogContent sx={{ maxHeight: "calc(100vh - 240px)" }}>
+            {searchValue && (
+              <>
+                <Box
+                  sx={{ pt: 0, pb: 1 }}
+                  display="flex"
+                  justifyContent="space-between"
                 >
-                  {searchValue}
-                </Typography>
-              </Typography>
-              <Link href="#" variant="body2" underline="hover">
-                Advanced search
-              </Link>
-            </Box>
-            <Divider sx={{ my: 1 }} />
+                  <Typography variant="body2" component="span">
+                    {stocksFinal ? (
+                      <>
+                        <Typography
+                          sx={{ fontWeight: "bold" }}
+                          variant="body1"
+                          component="span"
+                        >
+                          {count}
+                        </Typography>{" "}
+                        search results
+                      </>
+                    ) : (
+                      "Searching"
+                    )}{" "}
+                    for{" "}
+                    <Typography
+                      sx={{ fontWeight: "bold" }}
+                      variant="body1"
+                      component="span"
+                    >
+                      {searchValue}
+                    </Typography>
+                    {!stocksFinal && "…"}
+                  </Typography>
+                </Box>
+                <Divider />
+              </>
+            )}
+            {/* TODO: Make this a skeleton */}
             <List disablePadding>
-              <ListItem button>
-                <Hidden smDown>
-                  <ListItemAvatar>
-                    <Avatar
-                      sx={{
-                        background: (theme: Theme) =>
-                          theme.palette.secondary.main
-                      }}
-                    >
-                      <FindInPageIcon />
-                    </Avatar>
-                  </ListItemAvatar>
-                </Hidden>
-                <Box flex="1">
-                  <Box display="flex" justifyContent="space-between">
-                    <Link
-                      href="#"
-                      underline="hover"
-                      sx={{ fontWeight: 'bold' }}
-                      variant="body2"
-                    >
-                      Dashboard for Healthcare Platform
-                    </Link>
-                  </Box>
-                  <Typography
-                    component="span"
-                    variant="body2"
-                    sx={{
-                      color: (theme: Theme) =>
-                        lighten(theme.palette.secondary.main, 0.5)
-                    }}
-                  >
-                    This page contains all the necessary information for
-                    managing all hospital staff.
-                  </Typography>
-                </Box>
-                <ChevronRightIcon />
-              </ListItem>
-              <Divider sx={{ my: 1 }} component="li" />
-              <ListItem button>
-                <Hidden smDown>
-                  <ListItemAvatar>
-                    <Avatar
-                      sx={{
-                        background: (theme: Theme) =>
-                          theme.palette.secondary.main
-                      }}
-                    >
-                      <FindInPageIcon />
-                    </Avatar>
-                  </ListItemAvatar>
-                </Hidden>
-                <Box flex="1">
-                  <Box display="flex" justifyContent="space-between">
-                    <Link
-                      href="#"
-                      underline="hover"
-                      sx={{ fontWeight: 'bold' }}
-                      variant="body2"
-                    >
-                      Example Projects Application
-                    </Link>
-                  </Box>
-                  <Typography
-                    component="span"
-                    variant="body2"
-                    sx={{
-                      color: (theme: Theme) =>
-                        lighten(theme.palette.secondary.main, 0.5)
-                    }}
-                  >
-                    This is yet another search result pointing to a app page.
-                  </Typography>
-                </Box>
-                <ChevronRightIcon />
-              </ListItem>
-              <Divider sx={{ my: 1 }} component="li" />
-              <ListItem button>
-                <Hidden smDown>
-                  <ListItemAvatar>
-                    <Avatar
-                      sx={{
-                        background: (theme: Theme) =>
-                          theme.palette.secondary.main
-                      }}
-                    >
-                      <FindInPageIcon />
-                    </Avatar>
-                  </ListItemAvatar>
-                </Hidden>
-                <Box flex="1">
-                  <Box display="flex" justifyContent="space-between">
-                    <Link
-                      href="#"
-                      underline="hover"
-                      sx={{ fontWeight: 'bold' }}
-                      variant="body2"
-                    >
-                      Search Results Page
-                    </Link>
-                  </Box>
-                  <Typography
-                    component="span"
-                    variant="body2"
-                    sx={{
-                      color: (theme: Theme) =>
-                        lighten(theme.palette.secondary.main, 0.5)
-                    }}
-                  >
-                    Choose if you would like to show or not this typography
-                    section here...
-                  </Typography>
-                </Box>
-                <ChevronRightIcon />
-              </ListItem>
+              {stocksFinal
+                ? stocks.map((stock) => (
+                    <React.Fragment key={stock.ticker}>
+                      <HoverListItem
+                        style={{
+                          cursor: "pointer",
+                        }}
+                        onClick={() => {
+                          handleClose();
+                          window.open(`/#/stock/${stock.ticker}`, "_self");
+                        }}
+                      >
+                        <ListItemAvatar>
+                          <Avatar
+                            sx={{
+                              width: 80,
+                              height: 80,
+                              m: "-20px",
+                              background: "none",
+                            }}
+                            src={
+                              baseUrl +
+                              stockAPI +
+                              logoEndpoint +
+                              `/${stock.ticker}?dark=${
+                                theme.palette.mode === "dark"
+                              }`
+                            }
+                            alt=" "
+                          />
+                        </ListItemAvatar>
+                        <Box flex="1" sx={{ width: "calc(100% - 193px)" }}>
+                          <Typography variant="h5">{stock.name}</Typography>
+                          <Typography variant="body2" color="text.secondary">
+                            {stock.ticker}
+                          </Typography>
+                        </Box>
+                        <Typography sx={{ ml: 1 }} fontSize={18}>
+                          {emojiFlag(stock.country)}
+                        </Typography>
+                        <Box width={24} height={24} ml={1}>
+                          <SectorIcon
+                            industry={stock.industry}
+                            length={24}
+                            type={"Sector"}
+                          />
+                        </Box>
+                      </HoverListItem>
+                      <Divider component="li" />
+                    </React.Fragment>
+                  ))
+                : [...Array(count || 1)].map((_, index) => (
+                    <React.Fragment key={index}>
+                      <HoverListItem>
+                        <ListItemAvatar>
+                          <Skeleton variant="circular" width={40} height={40} />
+                        </ListItemAvatar>
+                        <Box flex="1" sx={{ width: "calc(100% - 193px)" }}>
+                          <Skeleton variant="text" width="100%" />
+                          <Skeleton variant="text" width={40} />
+                        </Box>
+                        <Skeleton
+                          variant="rectangular"
+                          width={24}
+                          height={24}
+                          sx={{ ml: 1 }}
+                        />
+                        <Skeleton
+                          variant="rectangular"
+                          width={24}
+                          height={24}
+                          sx={{ ml: 1 }}
+                        />
+                      </HoverListItem>
+                      <Divider component="li" />
+                    </React.Fragment>
+                  ))}
             </List>
-            <Divider sx={{ mt: 1, mb: 2 }} />
-            <Box sx={{ textAlign: 'center' }}>
-              <Button color="primary">View all search results</Button>
-            </Box>
           </DialogContent>
         )}
-      </DialogWrapper> */}
+      </DialogWrapper>
     </>
   );
 }
