+++ packages/rating-tracker-frontend/src/layouts/SidebarLayout/Header/Buttons/index.tsx
@@ -4,11 +4,11 @@ import HeaderNotifications from "./Notifications";
 
 function HeaderButtons() {
   return (
-    <Box sx={{ mr: 1 }}>
+    <Box sx={{ ml: 1 }}>
       <HeaderSearch />
-      <Box sx={{ mx: 0.5 }} component="span">
+      {/* <Box sx={{ mx: 0.5 }} component="span">
         <HeaderNotifications />
-      </Box>
+      </Box> */}
     </Box>
   );
 }
