+++ packages/rating-tracker-frontend/src/layouts/SidebarLayout/Header/index.tsx
@@ -13,7 +13,6 @@ import {
 } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
 import { SidebarContext } from "../../../contexts/SidebarContext";
-import CloseIcon from "@mui/icons-material/Close";
 
 import HeaderButtons from "./Buttons";
 import HeaderUserbox from "./Userbox";
@@ -23,7 +22,6 @@ const HeaderWrapper = styled(Box)(
   ({ theme }) => `
         height: ${theme.header.height};
         color: ${theme.header.textColor};
-        padding: ${theme.spacing(0, 2)};
         right: 0;
         z-index: 6;
         background-color: ${alpha(theme.header.background, 0.95)};
@@ -39,7 +37,7 @@ const HeaderWrapper = styled(Box)(
 );
 
 function Header() {
-  const { sidebarToggle, toggleSidebar } = useContext(SidebarContext);
+  const { toggleSidebar } = useContext(SidebarContext);
   const theme = useTheme();
 
   return (
@@ -47,6 +45,7 @@ function Header() {
       display="flex"
       alignItems="center"
       sx={{
+        px: 1,
         boxShadow:
           theme.palette.mode === "dark"
             ? `0 1px 0 ${alpha(
@@ -62,34 +61,31 @@ function Header() {
               )}`,
       }}
     >
-      <Stack
+      <Box
+        component="span"
+        sx={{
+          mr: 1,
+          display: "inline-block",
+          visibility: { lg: "hidden", xs: "visible" },
+        }}
+      >
+        <Tooltip arrow title="Toggle Menu">
+          <IconButton color="primary" onClick={toggleSidebar}>
+            <MenuIcon />
+          </IconButton>
+        </Tooltip>
+      </Box>
+      {/* <Stack
         direction="row"
         divider={<Divider orientation="vertical" flexItem />}
         alignItems="center"
         spacing={2}
       >
         <HeaderMenu />
-      </Stack>
+      </Stack> */}
       <Box display="flex" alignItems="center">
         <HeaderButtons />
         <HeaderUserbox />
-        <Box
-          component="span"
-          sx={{
-            ml: 2,
-            display: { lg: "none", xs: "inline-block" },
-          }}
-        >
-          <Tooltip arrow title="Toggle Menu">
-            <IconButton color="primary" onClick={toggleSidebar}>
-              {!sidebarToggle ? (
-                <MenuIcon fontSize="small" />
-              ) : (
-                <CloseIcon fontSize="small" />
-              )}
-            </IconButton>
-          </Tooltip>
-        </Box>
       </Box>
     </HeaderWrapper>
   );
